refactor: Refactor file indexing logic for improved clarity and reusability
